Fix project setup and imports
Module dependencies are placed in the module lib folder and
not in the common resources as before, avoiding namespace clashes
when some module uses another version of a library.

All project settings are stored in IDEAs format as .iml files, and
the old .classpath files for eclipse are ignored.

Recompiled the elang parser and lexer with the latest version of antlr.
The elang compiler is using the latest version as well.

The older version of antlr 4.0 is still used for the questioneer builder.
As there seem to be some incompativilities...

Make sure that old libs are removed before deplying this!!